Split Regions out of DataFile to separate files
